+++ asdf/commands/tests/test_exploded.py
@@ -5,7 +5,7 @@ import numpy as np
 import asdf
 from asdf import AsdfFile
 from asdf.commands import main
-from asdf.tests.helpers import assert_tree_match, get_file_sizes
+from asdf.tests._helpers import assert_tree_match, get_file_sizes
 
 
 def test_explode_then_implode(tmpdir):
