+++ asdf/commands/tests/test_info.py
@@ -3,7 +3,7 @@ from functools import partial
 import pytest
 
 from asdf.commands import main
-from asdf.tests import helpers
+from asdf.tests import _helpers as helpers
 
 from . import data as test_data
