+++ asdf/commands/tests/test_to_yaml.py
@@ -5,7 +5,7 @@ import numpy as np
 import asdf
 from asdf import AsdfFile
 from asdf.commands import main
-from asdf.tests.helpers import assert_tree_match, get_file_sizes
+from asdf.tests._helpers import assert_tree_match, get_file_sizes
 
 
 def test_to_yaml(tmpdir):
