+++ asdf/conftest.py
@@ -7,7 +7,7 @@ import pytest
 
 from asdf.tests.httpserver import HTTPServer, RangeHTTPServer
 
-collect_ignore = ["asdftypes.py", "fits_embed.py", "resolver.py", "type_index.py", "types.py"]
+collect_ignore = ["asdftypes.py", "fits_embed.py", "resolver.py", "type_index.py", "types.py", "tests/helpers.py"]
 
 
 @pytest.fixture()
