+++ asdf/tags/core/tests/test_ndarray.py
@@ -14,7 +14,7 @@ import asdf
 from asdf import util
 from asdf.exceptions import AsdfDeprecationWarning
 from asdf.tags.core import ndarray
-from asdf.tests import helpers
+from asdf.tests import _helpers as helpers
 from asdf.tests.objects import CustomTestType
 
 from . import data as test_data
