+++ asdf/tests/test_asdf.py
@@ -9,7 +9,7 @@ from asdf.entry_points import get_extensions
 from asdf.exceptions import AsdfWarning
 from asdf.extension import ExtensionManager, ExtensionProxy
 from asdf.extension._legacy import AsdfExtensionList
-from asdf.tests.helpers import assert_no_warnings, assert_tree_match, yaml_to_asdf
+from asdf.tests._helpers import assert_no_warnings, assert_tree_match, yaml_to_asdf
 from asdf.versioning import AsdfVersion
