+++ asdf/tests/test_helpers.py
@@ -2,7 +2,7 @@ import pytest
 
 from asdf import _types as types
 from asdf.exceptions import AsdfConversionWarning, AsdfDeprecationWarning, AsdfWarning
-from asdf.tests.helpers import assert_roundtrip_tree
+from asdf.tests._helpers import assert_roundtrip_tree
 
 
 def test_conversion_error(tmp_path):
