+++ asdf/tests/test_reference_files.py
@@ -6,7 +6,7 @@ import pytest
 from asdf import open as asdf_open
 from asdf import versioning
 
-from .helpers import assert_tree_match
+from ._helpers import assert_tree_match
 
 _REFFILE_PATH = os.path.join(os.path.dirname(__file__), "..", "..", "asdf-standard", "reference_files")
